rename TestNodeUtil to NodeTestUtil
