+++ hadoop-ozone/integration-test-recon/src/test/java/org/apache/hadoop/ozone/recon/TestReconAndAdminContainerCLI.java
@@ -63,7 +63,7 @@ import org.apache.hadoop.hdds.scm.container.ContainerManager;
 import org.apache.hadoop.hdds.scm.container.ReplicationManagerReport;
 import org.apache.hadoop.hdds.scm.container.replication.ReplicationManager;
 import org.apache.hadoop.hdds.scm.node.NodeManager;
-import org.apache.hadoop.hdds.scm.node.TestNodeUtil;
+import org.apache.hadoop.hdds.scm.node.NodeTestUtil;
 import org.apache.hadoop.hdds.scm.pipeline.Pipeline;
 import org.apache.hadoop.hdds.scm.pipeline.PipelineManager;
 import org.apache.hadoop.hdds.scm.server.StorageContainerManager;
@@ -225,7 +225,7 @@ class TestReconAndAdminContainerCLI {
 
     for (DatanodeDetails details : pipeline.getNodes()) {
       cluster.restartHddsDatanode(details, false);
-      TestNodeUtil.waitForDnToReachOpState(scmNodeManager, details, IN_SERVICE);
+      NodeTestUtil.waitForDnToReachOpState(scmNodeManager, details, IN_SERVICE);
     }
   }
 
@@ -254,19 +254,19 @@ class TestReconAndAdminContainerCLI {
     // First node goes offline.
     if (isMaintenance) {
       scmClient.startMaintenanceNodes(Collections.singletonList(
-          TestNodeUtil.getDNHostAndPort(nodeToGoOffline1)), 0, true);
+          NodeTestUtil.getDNHostAndPort(nodeToGoOffline1)), 0, true);
     } else {
       scmClient.decommissionNodes(Collections.singletonList(
-          TestNodeUtil.getDNHostAndPort(nodeToGoOffline1)), false);
+          NodeTestUtil.getDNHostAndPort(nodeToGoOffline1)), false);
     }
 
-    TestNodeUtil.waitForDnToReachOpState(scmNodeManager,
+    NodeTestUtil.waitForDnToReachOpState(scmNodeManager,
         nodeToGoOffline1, initialState);
 
     compareRMReportToReconResponse(underReplicatedState);
     compareRMReportToReconResponse(overReplicatedState);
 
-    TestNodeUtil.waitForDnToReachOpState(scmNodeManager,
+    NodeTestUtil.waitForDnToReachOpState(scmNodeManager,
         nodeToGoOffline1, finalState);
     // Every time a node goes into decommission,
     // a new replica-copy is made to another node.
@@ -281,19 +281,19 @@ class TestReconAndAdminContainerCLI {
     // Second node goes offline.
     if (isMaintenance) {
       scmClient.startMaintenanceNodes(Collections.singletonList(
-          TestNodeUtil.getDNHostAndPort(nodeToGoOffline2)), 0, true);
+          NodeTestUtil.getDNHostAndPort(nodeToGoOffline2)), 0, true);
     } else {
       scmClient.decommissionNodes(Collections.singletonList(
-          TestNodeUtil.getDNHostAndPort(nodeToGoOffline2)), false);
+          NodeTestUtil.getDNHostAndPort(nodeToGoOffline2)), false);
     }
 
-    TestNodeUtil.waitForDnToReachOpState(scmNodeManager,
+    NodeTestUtil.waitForDnToReachOpState(scmNodeManager,
         nodeToGoOffline2, initialState);
 
     compareRMReportToReconResponse(underReplicatedState);
     compareRMReportToReconResponse(overReplicatedState);
 
-    TestNodeUtil.waitForDnToReachOpState(scmNodeManager,
+    NodeTestUtil.waitForDnToReachOpState(scmNodeManager,
         nodeToGoOffline2, finalState);
 
     // There will be a replica copy for both maintenance and decommission.
@@ -305,16 +305,16 @@ class TestReconAndAdminContainerCLI {
     compareRMReportToReconResponse(overReplicatedState);
 
     scmClient.recommissionNodes(Arrays.asList(
-        TestNodeUtil.getDNHostAndPort(nodeToGoOffline1),
-        TestNodeUtil.getDNHostAndPort(nodeToGoOffline2)));
+        NodeTestUtil.getDNHostAndPort(nodeToGoOffline1),
+        NodeTestUtil.getDNHostAndPort(nodeToGoOffline2)));
 
-    TestNodeUtil.waitForDnToReachOpState(scmNodeManager,
+    NodeTestUtil.waitForDnToReachOpState(scmNodeManager,
         nodeToGoOffline1, IN_SERVICE);
-    TestNodeUtil.waitForDnToReachOpState(scmNodeManager,
+    NodeTestUtil.waitForDnToReachOpState(scmNodeManager,
         nodeToGoOffline2, IN_SERVICE);
 
-    TestNodeUtil.waitForDnToReachPersistedOpState(nodeToGoOffline1, IN_SERVICE);
-    TestNodeUtil.waitForDnToReachPersistedOpState(nodeToGoOffline2, IN_SERVICE);
+    NodeTestUtil.waitForDnToReachPersistedOpState(nodeToGoOffline1, IN_SERVICE);
+    NodeTestUtil.waitForDnToReachPersistedOpState(nodeToGoOffline2, IN_SERVICE);
 
     compareRMReportToReconResponse(underReplicatedState);
     compareRMReportToReconResponse(overReplicatedState);
